feat(blog-web;blog): AI文档暂时使用Fixture PDF做Mock测试

+++ blog/module-ai/src/main/java/top/fxmarkbrown/blog/service/impl/AiArticleToolServiceImpl.java
@@ -133,7 +133,7 @@ public class AiArticleToolServiceImpl implements AiArticleToolService {
             if (item == null || item.getArticleId() == null) {
                 continue;
             }
-            Long articleId = item.getArticleId().longValue();
+            Long articleId = item.getArticleId();
             if (!seenArticleIds.add(articleId)) {
                 continue;
             }
